+++ CHANGELOG.md
@@ -7,9 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.2] - 2026-01-19
 
-### Added
+### Fixed
 
 - Fix Packaging. The VIX didn't package dependencies.
+- Filter out deprecated models flagged by models.dev status.
 
 ## [0.1.1] - 2026-01-17